quinto commit

+++ src/main/java/com/javanautas/fake_api_us/business/services/ProdutoService.java
@@ -3,9 +3,10 @@ package com.javanautas.fake_api_us.business.services;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import com.javanautas.fake_api_us.apiv1.dto.ProductsDto;
 import com.javanautas.fake_api_us.business.converter.ProdutoConverter;
+import com.javanautas.fake_api_us.dto.ProdutoDto;
 import com.javanautas.fake_api_us.infrastructure.entities.ProdutoEntity;
 import com.javanautas.fake_api_us.infrastructure.repositories.ProdutoRepository;
 
@@ -20,7 +21,7 @@ public class ProdutoService {
 	private final ProdutoRepository produtoRepository;
 	private final ProdutoConverter produtoConverter;
 
-	public ProdutoEntity salvarProdutos(ProdutoEntity produtoEntity) {
+	public ProdutoEntity salvarProduto(ProdutoEntity produtoEntity) {
 		try {
 			return produtoRepository.save(produtoEntity);
 		} catch (Exception e) {
@@ -30,7 +31,7 @@ public class ProdutoService {
 	}
 	
 	
-	public ProductsDto salvarProdutoDto(ProductsDto produtoDto) {
+	public ProdutoDto salvarProdutoDto(ProdutoDto produtoDto) {
 		try {
 			ProdutoEntity produtoEntity = produtoConverter.ProductToEntity(produtoDto);
 			return produtoConverter.ProductToDto(produtoRepository.save(produtoEntity));
@@ -50,7 +51,7 @@ public class ProdutoService {
 //	}
 	
 	
-	public ProductsDto buscarProdutoPorNome(String produtoNome) {
+	public ProdutoDto buscarProdutoPorNome(String produtoNome) {
 		try {
 			return produtoConverter.ProductToDto(produtoRepository.findByNome(produtoNome));
 		} catch (Exception e) {
@@ -59,7 +60,7 @@ public class ProdutoService {
 	}
 	
 	
-	public List<ProductsDto> buscaTodosProdutos() {
+	public List<ProdutoDto> buscarTodosProdutos() {
 		try {
 			return produtoConverter.ProductToListDto(produtoRepository.findAll());
 		} catch (Exception e) {
@@ -68,9 +69,15 @@ public class ProdutoService {
 	}
 	
 	
-	public void deletaProduto(String produtoNome) {
+	@Transactional
+	public void deletarProduto(String produtoNome) {
 		try {
-			produtoRepository.deleteByNome(produtoNome);
+			if (existsPorNome(produtoNome)) { 
+			 produtoRepository.deleteByNome(produtoNome);
+			}else {
+	            throw new RuntimeException("Produto não encontrado: " + produtoNome);
+	        }
+					
 		} catch (Exception e) {
 			throw new RuntimeException(format("Erro ao deletar produto por nome",produtoNome),e);   
 		}
@@ -86,10 +93,10 @@ public class ProdutoService {
     }
 	
 	
-	public ProductsDto updateProduto(String id, ProductsDto produtoDto) {
+	public ProdutoDto atualizarProduto(String id, ProdutoDto produtoDto) {
 		try {
 			ProdutoEntity produtoEntity = produtoRepository.findById(id).orElseThrow(()->new RuntimeException("Id não existe no banco"));
-			salvarProdutos(produtoConverter.ProductToEntityUpdate(produtoEntity,produtoDto,id));
+			salvarProduto(produtoConverter.ProductToEntityUpdate(produtoEntity,produtoDto,id));
 			return produtoConverter.ProductToDto(produtoRepository.findByNome(produtoEntity.getNome()));
 		} catch (Exception e) {
 			 throw new RuntimeException(format("Erro ao atualizar produto"));
